Implement GenerationLoader and add stub for Mapper
The mapper is not implemented yet.
It will map the json representation (a raw POCO) to the model.
