+++ NEAT_Visualizer/Business/DataLoaders/JsonToModelMapper.cs
@@ -1,0 +1,12 @@
+﻿using NEAT_Visualizer.Model;
+
+namespace NEAT_Visualizer.Business.DataLoaders
+{
+  public static class JsonToModelMapper
+  {
+    public static Generation ToModel(this JsonRepresentation.Rootobject jsonRoot)
+    {
+      throw new System.NotImplementedException();
+    }
+  }
+}
