+++ NEAT_Visualizer/Model/Generation.cs
@@ -4,6 +4,17 @@ namespace NEAT_Visualizer.Model
 {
   public class Generation
   {
-    public IList<Species> Species { get; set; } = new List<Species>();
+    public Generation(uint generationsPassed, uint generationSize, IList<Species> species)
+    {
+      GenerationsPassed = generationsPassed;
+      GenerationSize = generationSize;
+      Species = species;
+    }
+
+    public uint GenerationSize { get; set; }
+
+    public uint GenerationsPassed { get; set; }
+
+    public IList<Species> Species { get; set; }
   }
 }
